test(repl): move NewServerID tests from client_test.go to repl_test.go
NewServerID was moved to repl.go; co-locate its tests in repl_test.go.
TestNewServerIDConcurrent (birthday-paradox-tolerant concurrent generation
of 10k IDs) and TestNewServerIDRange (range-bound check across 1000
IDs) move unchanged.
